requirements.txt
imutils
dlib-bin
numpy
opencv-python
scipy
requests
pillow
librosa==0.10.0.post2
opencv-contrib-python
tqdm
numba
imutils
imageio_ffmpeg
git+https://github.com/suno-ai/bark.git
insightface==0.7.3
onnx==1.14.0
onnxruntime==1.15.0
onnxruntime-gpu==1.15.0
opencv-python>=4.8.0
ifnude